+++ WhateverGreen/kern_weg.cpp
@@ -237,6 +237,13 @@ void WEG::processKernel(KernelPatcher &patcher) {
 			if (appleBacklightPatch == APPLBKL_DETECT && devInfo->videoBuiltin != nullptr)
 				WIOKit::getOSDataValue(devInfo->videoBuiltin, "applbkl", appleBacklightPatch);
 
+			if (appleBacklightCustomName == nullptr && devInfo->videoBuiltin != nullptr) {
+				appleBacklightCustomName = OSDynamicCast(OSData, devInfo->videoBuiltin->getProperty("applbkl-name"));
+				appleBacklightCustomData = OSDynamicCast(OSData, devInfo->videoBuiltin->getProperty("applbkl-data"));
+				if (appleBacklightCustomName == nullptr || appleBacklightCustomData == nullptr)
+					appleBacklightCustomName = appleBacklightCustomData = nullptr;
+			}
+
 			for (size_t i = 0; i < extNum; i++) {
 				auto &v = devInfo->videoExternal[i];
 				processExternalProperties(v.video, devInfo, v.vendor);
@@ -247,6 +254,13 @@ void WEG::processKernel(KernelPatcher &patcher) {
 
 				if (appleBacklightPatch == APPLBKL_DETECT)
 					WIOKit::getOSDataValue(v.video, "applbkl", appleBacklightPatch);
+
+				if (appleBacklightCustomName == nullptr) {
+					appleBacklightCustomName = OSDynamicCast(OSData, v.video->getProperty("applbkl-name"));
+					appleBacklightCustomData = OSDynamicCast(OSData, v.video->getProperty("applbkl-data"));
+					if (appleBacklightCustomName == nullptr || appleBacklightCustomData == nullptr)
+						appleBacklightCustomName = appleBacklightCustomData = nullptr;
+				}
 			}
 
 			// Note, disabled Optimus will make videoExternal 0, so this case checks for active IGPU only.
@@ -754,13 +768,26 @@ bool WEG::wrapApplePanelSetDisplay(IOService *that, IODisplay *display) {
 			panels = OSDynamicCast(OSDictionary, rawPanels);
 
 			if (panels) {
+				const char *customName = nullptr;
+				if (callbackWEG->appleBacklightCustomName != nullptr) {
+					auto length = callbackWEG->appleBacklightCustomName->getLength();
+					const char *customNameBytes = static_cast<const char *>(callbackWEG->appleBacklightCustomName->getBytesNoCopy());
+					if (length > 0 && customNameBytes[length - 1] == '\0')
+						customName = customNameBytes;
+				}
+
 				for (auto &entry : appleBacklightData) {
-					auto pd = OSData::withBytes(entry.deviceData, sizeof(entry.deviceData));
-					if (pd) {
-						panels->setObject(entry.deviceName, pd);
-						// No release required by current AppleBacklight implementation.
+					if (customName != nullptr && strcmp(customName, entry.deviceName) == 0) {
+						panels->setObject(entry.deviceName, callbackWEG->appleBacklightCustomData);
+						DBGLOG("weg", "using custom panel data for %s device", entry.deviceName);
 					} else {
-						SYSLOG("weg", "panel start cannot allocate %s data", entry.deviceName);
+						auto pd = OSData::withBytes(entry.deviceData, sizeof(entry.deviceData));
+						if (pd) {
+							panels->setObject(entry.deviceName, pd);
+							// No release required by current AppleBacklight implementation.
+						} else {
+							SYSLOG("weg", "panel start cannot allocate %s data", entry.deviceName);
+						}
 					}
 				}
 				that->setProperty("ApplePanels", panels);